ColliderMap: Added find helper method to ColliderMap
Added remove collider function to collider map, this required a rewrite of how collider IDs are handled. Remove uses find.
Now, a collider with an id != -1 may be added to a ColliderMap if you pass the true to the new force parameter. More info in the method descriptor.
Check colliders has a new precondition, the col param must be in the ColliderMap, and will return null if the precondition is not met.

TestRunnerTwoLoop:
Added a hold to noclip button to, bound to the F key.